idris.yml
AMP keeps users within Google’s domain and diverts traffic away from other websites in the pretext of benefit to end users. At a scale of billions of users, this has the effect of further reinforcing Google’s dominance of the Web. Publishers should not be compelled by Google’s search dominance to put their content under a Google umbrella.
The Web is not Google, and should not be just Google.